# ikeyit-classroom/ikeyit-classroom-interfaces-api/ikeyit-classroom-openapi.yml
openapi: 3.0.3
info:
  title: ikeyit Classroom API
  description: |
    This is the API documentation for the ikeyit Classroom system.
    It provides endpoints for managing courses, teachers, students, class schedules, and enrollments.
  version: 1.0.0
servers:
  - url: http://localhost:6111/api/classroom/v1
    description: Current server
...
paths:
  /students/me:
    post:
      tags:
        - students
      summary: Register student information
      description: Register current user as a student
      security:
        - oauth2: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStudentCMD'
      responses:
        '200':
          description: Student registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentDTO'
    put:
      tags:
        - students
      summary: Update student information
      description: Update current user's student information
      security:
        - oauth2: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStudentCMD'
      responses:
        '200':
          description: Student information updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentDTO'
    get:
      tags:
        - students
      summary: Get student information
      description: Get current user's student information
      security:
        - oauth2: []
      responses:
        '200':
          description: Student information retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentDTO'
  /students/{userId}:
    get:
      tags:
        - students
      summary: Get student by user ID
      description: Get student information by user ID
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Student information retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentDTO'
  /teachers/me:
    post:
      tags:
        - teachers
      summary: Register teacher information
      description: Register current user as a teacher
      security:
        - oauth2: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTeacherCMD'
      responses:
        '200':
          description: Teacher registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeacherDTO'
    put:
      tags:
        - teachers
      summary: Update teacher information
      description: Update current user's teacher information
      security:
        - oauth2: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTeacherCMD'
      responses:
        '200':
          description: Teacher information updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeacherDTO'
    get:
      tags:
        - teachers
      summary: Get teacher information
      description: Get current user's teacher information
      security:
        - oauth2: []
      responses:
        '200':
          description: Teacher information retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeacherDTO'
  /teachers/{teacherUserId}:
    get:
      tags:
        - teachers
      summary: Get teacher by user ID
      description: Get teacher information by user ID
      parameters:
        - name: teacherUserId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Teacher information retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeacherDTO'
...
components:
  securitySchemes:
    oauth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: /oauth2/authorize
          tokenUrl: /oauth2/token
          scopes: {}
  schemas:
    CreateStudentCMD:
      type: object
      properties:
        userId:
          type: integer
          format: int64
          description: User ID of the student
        name:
          type: string
          description: Name of the student
        studentNo:
          type: string
          description: Student number
        faculty:
          type: string
          description: Faculty of the student
        className:
          type: string
          description: Class name of the student
    UpdateStudentCMD:
      type: object
      properties:
        operatorId:
          type: integer
          format: int64
          description: ID of the user performing the operation
        name:
          type: string
          description: Name of the student
        faculty:
          type: string
          description: Faculty of the student
        className:
          type: string
          description: Class name of the student
    StudentDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: ID of the student
        userId:
          type: integer
          format: int64
          description: User ID of the student
        name:
          type: string
          description: Name of the student
        studentNo:
          type: string
          description: Student number
        faculty:
          type: string
          description: Faculty of the student
        className:
          type: string
          description: Class name of the student
        createdAt:
          type: string
          format: date-time
          description: Creation time
        updatedAt:
          type: string
          format: date-time
          description: Last update time
    CreateTeacherCMD:
      type: object
      properties:
        userId:
          type: integer
          format: int64
          description: User ID of the teacher
        name:
          type: string
          description: Name of the teacher
        faculty:
          type: string
          description: Faculty of the teacher
        title:
          type: string
          description: Title of the teacher
    UpdateTeacherCMD:
      type: object
      properties:
        operatorId:
          type: integer
          format: int64
          description: ID of the user performing the operation
        name:
          type: string
          description: Name of the teacher
        faculty:
          type: string
          description: Faculty of the teacher
        title:
          type: string
          description: Title of the teacher
    TeacherDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: ID of the teacher
        userId:
          type: integer
          format: int64
          description: User ID of the teacher
        name:
          type: string
          description: Name of the teacher
        faculty:
          type: string
          description: Faculty of the teacher
        title:
          type: string
          description: Title of the teacher
        createdAt:
          type: string
          format: date-time
          description: Creation time
        updatedAt:
          type: string
          format: date-time
          description: Last update time
...